Making stackline setup a little more resilient

+++ non-linked-scripts/setup-stackline.sh
@@ -1,8 +1,17 @@
 #!/bin/bash -e
 
 # Just attempt to replicate what https://github.com/Hammerspoon/hammerspoon/blob/da2f41aa2c008c84a746ffc0d5a0d3291d6e5264/extensions/ipc/init.lua#L266 does.
-ln -s /Applications/Hammerspoon.app/Contents/Resources/extensions/hs/ipc/bin/hs "$HOME/.local/bin/hs"
-ln -s /Applications/Hammerspoon.app/Contents/Resources/extensions/hs/ipc/share/man/man1/hs.1 /opt/homebrew/share/man/man1/hs.1
 
+if [ ! \( -e "$HOME/.local/bin/hs" \) ]
+then
+  ln -s /Applications/Hammerspoon.app/Contents/Resources/extensions/hs/ipc/bin/hs "$HOME/.local/bin/hs"
+fi
 
-sed 's/\/usr\/local\/bin\/jq/\/opt\/homebrew\/bin\/jq/' "$HOME/.hammerspoon/stackline/conf.lua" | sed 's/\/usr\/local\/bin\/yabai/\/opt\/homebrew\/bin\/yabai/'> "$HOME/.hammerspoon/stackline/conf.lua"
+if [ ! \( -e "/opt/homebrew/share/man/man1/hs.1" \) ]
+then
+  ln -s /Applications/Hammerspoon.app/Contents/Resources/extensions/hs/ipc/share/man/man1/hs.1 /opt/homebrew/share/man/man1/hs.1
+fi
+
+sed 's/\/usr\/local\/bin\/jq/\/opt\/homebrew\/bin\/jq/' "$HOME/.hammerspoon/stackline/conf.lua" | sed 's/\/usr\/local\/bin\/yabai/\/opt\/homebrew\/bin\/yabai/' > "$HOME/.hammerspoon/stackline/temp-config.lua"
+
+mv "$HOME/.hammerspoon/stackline/temp-config.lua" "$HOME/.hammerspoon/stackline/conf.lua"
